+++ lib/runtime/command/command-dispatch.js
@@ -2,8 +2,8 @@
 'use strict';
 
 const { Broadcast, Logger } = require('ranvier');
-const { parseInput } = require('../parse-input');
-const Mutator = require('./mutator');
+const { parseInput } = require('../../parse-input');
+const Mutator = require('../mutation/mutator');
 const EntityResolution = require('./entity-resolution');
 const RenderDispatch = require('./render-dispatch');
 
@@ -58,7 +58,7 @@ const RenderDispatch = require('./render-dispatch');
 
 /** @typedef {import('ranvier/types/GameState')} GameState */
 /** @typedef {import('ranvier/types/Command')} Command */
-/** @typedef {import('./mutator').MutationPlan} MutationPlan */
+/** @typedef {import('../mutation/mutator').MutationPlan} MutationPlan */
 
 /**
  * @typedef {object} Session
