+++ lib/runtime/command/entity-resolution.js
@@ -1,8 +1,8 @@
 // @ts-check
 'use strict';
 
-const Helper = require('../helpers/entity-resolution-helper');
-const GoExitHookHelper = require('../doors/go-exit-hook-helper');
+const Helper = require('../../helpers/entity-resolution-helper');
+const GoExitHookHelper = require('../../doors/go-exit-hook-helper');
 const Syntax = require('./verb-local-syntax');
 
 /**
