+++ lib/runtime/command/semantic-message.js
@@ -4,7 +4,7 @@
 /**
  * Standalone semantic-message renderer.
  *
- * This module is wired through `lib/session/render-dispatch.js` and is also
+ * This module is wired through `lib/runtime/command/render-dispatch.js` and is also
  * used directly by semantic-message focused tests.
  *
  * Design notes:
