+++ lib/runtime/command/verb-local-syntax.js
@@ -1,7 +1,7 @@
 // @ts-check
 'use strict';
 
-const Helper = require('../helpers/entity-resolution-helper');
+const Helper = require('../../helpers/entity-resolution-helper');
 
 const SLOT_KINDS = new Set(['TEXT', 'WORD', 'NUMBER', 'ENTITY', 'LIVING', 'EXIT']);
 const VARIABLE_WIDTH_SLOTS = new Set(['TEXT', 'ENTITY', 'LIVING', 'EXIT']);
