+++ lib/runtime/conversation/conversation-definition-service.js
@@ -4,7 +4,7 @@
 const fs = require('fs');
 const path = require('path');
 const { Data, Logger } = require('ranvier');
-const { deepFreeze } = require('../helpers/deep-freeze');
+const { deepFreeze } = require('../../helpers/deep-freeze');
 const { evaluateConversationRuntime } = require('./conversation-runtime');
 const { validateConversationDefinition } = require('./conversation-definition-validation');
