+++ lib/runtime/conversation/conversation-definition-validation.js
