+++ lib/runtime/conversation/conversation-state.js
@@ -1,7 +1,7 @@
 // @ts-check
 'use strict';
 
-const { getPlayerMetadata, parsePath } = require('./player-metadata');
+const { getPlayerMetadata, parsePath } = require('../mutation/player-metadata');
 
 const CONVERSATION_METADATA_ROOT = 'conversations';
