+++ lib/runtime/mutation/mutator.js
@@ -3,7 +3,7 @@
 
 const util = require('util');
 const { Broadcast, Logger } = require('ranvier');
-const { getVirtualDoorService } = require('../doors/virtual-door-service');
+const { getVirtualDoorService } = require('../../doors/virtual-door-service');
 const { parsePath } = require('./player-metadata');
 const { deleteWorldMetadataPath, getOrCreateWorldMetadataValuesRoot } = require('./world-metadata-service');
